added author

+++ src/com/hackerrank/problemsolving/JavaAnagrams.java
@@ -1,5 +1,11 @@
 package com.hackerrank.problemsolving;
 
+/**
+ * 
+ * @author tabishmatin
+ *
+ */
+
 public class JavaAnagrams {
 
     static boolean isAnagram(String s1, String s2) {
@@ -10,11 +16,11 @@ public class JavaAnagrams {
         if(s1.length()==s2.length())
 
         {
-            int[] a = new int[256];
+            int[] a = new int[256]; 
             int[] b = new int[256];
             for (int i = 0; i < s1.length(); i++) {
-                a[(int) s1.charAt(i)] += 1;
-                b[(int) s2.charAt(i)] += 1;
+                a[s1.charAt(i)] += 1;
+                b[s2.charAt(i)] += 1;
             }
             for (int i = 0; i < 256; i++) {
                 if (a[i] != b[i])
